large commit to add recent code updates, sample dataset, and scripts to build and run sample experiment
Note that reduce PA_90_weighted_2-10.shuf size by using the first 5M edges only.

+++ source/src/main/java/com/googlecode/javaewah/datastructure/BitSet.java
@@ -22,7 +22,7 @@ public final class BitSet implements Cloneable {
         if (sizeinbits < 0) {
             throw new NegativeArraySizeException("negative number of bits: " + sizeinbits);
         }
-        this.data = new long[(sizeinbits + 63) / 64];
+        this.data = new long[(int) (((double)sizeinbits + 63) / 64)];
     }
 
     public int andcardinality(BitSet bs) {
